+++ src/Aks/Aks/ChangeLog.md
@@ -21,6 +21,7 @@
 * Bumped API version to 2022-09-01
 * Added parameter `NodeOsSKU` for `New-AzAksCluster` and parameter `OsSKU` for `New-AzAksNodePool`
 * Added parameter `Mode` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added property `NodeImageVersion` for the output of `Get-AzAksNodePool`[#19893]
 
 ## Version 5.0.1
 * Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]